+++ app/Http/Livewire/Style/Table.php
@@ -9,10 +9,17 @@ use Livewire\WithPagination;
 class Table extends Component
 {
     use WithPagination;
+    public $search = '';
+    
+    public function updating($name, $value){
+        $this->resetPage();
+    } 
+
     public function render()
     {
+        $styles = Style::where('name', 'like', '%'.$this->search.'%');
         return view('livewire.style.table', [
-            'collection' => Style::paginate(10)
+            'collection' => $styles->paginate(10)
         ]);
     }
 }