+++ resources/views/components/ui/link.blade.php
@@ -1,0 +1,6 @@
+<div>
+    <!-- Breathing in, I calm body and mind. Breathing out, I smile. - Thich Nhat Hanh -->
+    <a href="{{ $route }}" class="{{ $css ?? 'text-indigo-500' }} hover:text-indigo-800">
+        {{ $slot }}
+    </a>
+</div>